home_3 main presenter unused and removed one interface for all screens in Screens.kt
