--- app/src/main/java/com/somethingsimple/poplibs/ui/main/MainView.kt
@@ -1,8 +1,0 @@
-package com.somethingsimple.poplibs.ui.main
-
-import moxy.MvpView
-import moxy.viewstate.strategy.AddToEndSingleStrategy
-import moxy.viewstate.strategy.StateStrategyType
-
-@StateStrategyType(AddToEndSingleStrategy::class)
-interface MainView : MvpView
