+++ app/src/main/java/com/somethingsimple/poplibs/ui/users/Screens.kt
@@ -2,11 +2,16 @@ package com.somethingsimple.poplibs.ui.users
 
 import com.github.terrakok.cicerone.Screen
 import com.github.terrakok.cicerone.androidx.FragmentScreen
+import com.somethingsimple.poplibs.data.model.GithubUser
+import com.somethingsimple.poplibs.ui.users.detail.UserDetailFragment
 
 interface IScreens {
     fun users(): Screen
+    fun userDetails(user: GithubUser): Screen
 }
 
-class AndroidScreens : IScreens {
+object PopLibsAppScreens : IScreens {
     override fun users() = FragmentScreen { UsersFragment.newInstance() }
+    override fun userDetails(user: GithubUser): Screen =
+        FragmentScreen { UserDetailFragment.newInstance(user) }
 }
